backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3            # via black
aspy.yaml==1.1.2          # via pre-commit
attrs==18.2.0             # via black
black==18.9b0
cfgv==1.4.0               # via pre-commit
click==7.0                # via black, pip-tools
coverage==4.5.2
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.5
django==2.1.5
djangorestframework==3.9.1
factory-boy==2.11.1
faker==1.0.2              # via factory-boy
identify==1.2.1           # via pre-commit
importlib-metadata==0.8   # via pre-commit
importlib-resources==1.0.2  # via pre-commit
isort==4.3.9
nodeenv==1.3.3            # via pre-commit
pillow==5.4.1
pip-tools==3.3.2
pipdeptree==0.13.2
pre-commit==1.14.3
python-dateutil==2.8.0    # via faker
pytz==2018.9
pyyaml==3.13              # via aspy.yaml, pre-commit
six==1.12.0               # via cfgv, django-extensions, faker, pip-tools, pre-commit, python-dateutil
sqlparse==0.2.4           # via django-debug-toolbar
text-unidecode==1.2       # via faker
toml==0.10.0              # via black, pre-commit
virtualenv==16.3.0        # via pre-commit
zipp==0.3.3               # via importlib-metadata
